+++ calcipy/cli/__init__.py
@@ -1,0 +1,5 @@
+"""Calcipy Command Line Interface."""
+
+# PLANNED: Take a look at a few useful issues for cement:
+# FAQ: The right decision to layout your Cement 3 CLI app (and test)
+# 	https://github.com/datafolklabs/cement/issues/602