+++ docs/docs/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Feat
 
+- add cct command
+- initialize the cement CLI application
 - replace toml with tomli (#74)
 - move wily to nox to reduce version conflicts
 - add pip-audit
@@ -9,10 +11,12 @@
 
 ### Fix
 
+- cct was not being created as an alias. Use full command
 - wily
 
 ### Refactor
 
+- move file_search to top-level
 - just run pre-commit install
 - replace pass with ... to keep it from being removed